# leoWIfi/public/swagger/api.json
{"swagger":"2.0","info":{"title":"LeoWifi API","description":"# 访问规则\n * DMAIN+APP_ID+API_VERSION+API_NAME ex:http://api.leomaster.com/wifi/v1/init\n * 加密方式待定\n<pre>**header** </pre>\n","version":"1.0"},"basePath":"/wifi/v1","tags":[{"name":"wifi","description":"wifi相关API"}],"schemes":["http"],"consumes":["application/json","multipart/form-data"],"produces":["application/json"],"paths":{"/wifi":{"post":{"tags":["wifi"],"description":"采集wifi信息规则:  \n* 反查国家城市信息IP优先级 wifi连接时的IP>上报IP 如没有查到则置为空.  \n* 目前不保存无密码可直连的wifi信息.  \n\n\n 规则分类|处理方式|规则描述 \n------|------|------ \n_id|upsert|tryTime>=系统最后记录状态时间 \nbssid|upsert|同一国家,城市,bssid一致.并且tryTime>=系统最后记录状态时间 \n其他|insert|直接更新 \n","summary":"采集wifi信息","parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["device_id","infos"],"properties":{"device_id":{"type":"string","description":"客户端设备ID"},"infos":{"type":"array","items":{"$ref":"#/definitions/wifiInfoGather"}}}}}],"responses":{"200":{"description":"采集wifi信息","schema":{"type":"object","properties":{"code":{"type":"number","description":"error code","default":0},"msg":{"type":"string","description":"error message"},"data":{"type":"object","properties":{"id":{"type":"string","description":""}}}}},"examples":{"application/json":{"code":0,"msg":"","data":[]}}},"x-error":{"schema":{"required":["type"],"type":"object","$ref":"#/definitions/ErrorRtn"}}}}},"/hotspot":{"post":{"tags":["wifi"],"description":"采集wifi热点信息规则（暂定）:  \n* 反查国家城市信息IP优先级 wifi连接时的IP>上报IP 如没有查到则置为空.  \n\n\n 规则分类|处理方式|规则描述 \n------|------|------ \n_id|upsert|tryTime>=系统最后记录状态时间 \nbssid|upsert|同一国家,城市,bssid一致.并且tryTime>=系统最后记录状态时间 \n其他|insert|直接更新 \n","summary":"采集wifi热点信息（暂定）","parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["device_id","infos"],"properties":{"device_id":{"type":"string","description":"客户端设备ID"},"infos":{"type":"array","items":{"$ref":"#/definitions/wifiInfo"}}}}}],"responses":{"200":{"description":"采集wifi信息","schema":{"type":"object","properties":{"code":{"type":"number","description":"error code","default":0},"msg":{"type":"string","description":"error message"},"data":{"type":"object","properties":{"_id":{"type":"string","description":"数据库Id"}}}}},"examples":{"application/json":{"code":0,"msg":"","data":[]}}},"x-error":{"schema":{"required":["type"],"type":"object","$ref":"#/definitions/ErrorRtn"}}}}},"/findwifi":{"post":{"tags":["wifi"],"description":"挖掘wifi信息规则:  \n* 仅返回有密码并且状态可连接的wifi信息  \n* 匹配规则待改进.  \n\n\n 规则分类|规则描述 \n------|------ \n_id|_id匹配并且connectable=true \nbssid|同一国家,城市,bssid.并且connetable=true \nssid|同一国家,城市,ssid. connectable=true,并且以请求经纬度为中心半径500米距离查找 \n","summary":"挖掘wifi信息","parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["device_id","infos"],"properties":{"device_id":{"type":"string","description":"客户端设备ID"},"latitude":{"type":"string","description":"纬度"},"longitude":{"type":"string","description":"经度"},"infos":{"type":"array","items":{"$ref":"#/definitions/simpleWifiInfo"}}}}}],"responses":{"200":{"description":"挖掘wifi信息","schema":{"type":"object","properties":{"code":{"type":"number","description":"error code","default":0},"msg":{"type":"string","description":"error message"},"data":{"type":"object","properties":{"infos":{"type":"array","items":{"$ref":"#/definitions/wifiInfoResponse"}}}}}},"examples":{"application/json":{"err":0,"msg":"","data":{"_id":""}}}},"x-error":{"schema":{"required":["type"],"type":"object","$ref":"#/definitions/ErrorRtn"}}}}},"/config":{"get":{"tags":["wifi"],"description":"获取客户端设置","summary":"获取客户端设置,目前以配置文件的形式保存在客户端","responses":{"200":{"description":"客户端设置推送","schema":{"type":"object","properties":{"code":{"type":"number","description":"error code","default":0},"msg":{"type":"string","description":"error message"},"data":{"type":"object","properties":{"enableWifiCollect":{"type":"boolean","description":"是否允许采集","default":true},"showFreeWifiCount":{"type":"number","description":"显示免费wifi的数量","default":5},"gatherWifiCountOnce":{"type":"number","description":"每次采集wifi信息上传记录数","default":10},"gatherNSWifi":{"type":"boolean","description":"采集不允许分享的wifi密码","default":true}}}}},"examples":{"application/json":{"code":0,"msg":"","data":{"enableWifiCollect":true,"showFreeWifiCount":5,"gatherWifiCountOnce":10,"gatherNSWifi":true}}}},"x-error":{"schema":{"required":["type"],"type":"object","$ref":"#/definitions/ErrorRtn"}}}}},"/hotspot/poster":{"post":{"tags":["wifi"],"description":"采集wifi信息规则:  \n* 反查国家城市信息IP优先级 wifi连接时的IP>上报IP 如没有查到则置为空.  \n* 目前不保存无密码可直连的wifi信息.  \n\n\n 规则分类|处理方式|规则描述 \n------|------|------ \n_id|upsert|tryTime>=系统最后记录状态时间 \nbssid|upsert|同一国家,城市,bssid一致.并且tryTime>=系统最后记录状态时间 \n其他|insert|直接更新 \n","summary":"上传热点头像","parameters":[{"name":"avatar","type":"file","in":"formData","required":true,"description":"头像文件"},{"name":"_id","type":"string","in":"formData","description":"数据库ID"}],"responses":{"200":{"description":"上传热点头像","schema":{"type":"object","properties":{"code":{"type":"number","description":"error code","default":0},"msg":{"type":"string","description":"error message"},"data":{"type":"object","properties":{"id":{"type":"string","description":""}}}}},"examples":{"application/json":{"code":0,"msg":"","data":[]}}},"x-error":{"schema":{"required":["type"],"type":"object","$ref":"#/definitions/ErrorRtn"}}}}},"/hotspot/poster/{name}":{"get":{"tags":["wifi"],"description":"根据名称获取热点海报","summary":"","parameters":[{"name":"name","type":"string","in":"path","required":true,"description":"海报名称"}],"responses":{"200":{"description":"根据名称获取热点海报"},"x-error":{"schema":{"required":["type"],"type":"object","$ref":"#/definitions/ErrorRtn"}}}}}},"definitions":{"wifiInfoGather":{"type":"object","properties":{"_id":{"type":"string","description":"db objectId"},"ssid":{"type":"string","description":"wifi ssid"},"bssid":{"type":"string","description":"wifi bssid"},"level":{"type":"integer","description":"wifi level"},"sec_level":{"type":"integer","description":"wifi sec_level"},"capabilities":{"type":"string","description":"wifi capabilities"},"frequency":{"type":"integer","description":"wifi frequency"},"password":{"type":"string","description":"wifi password"},"identity":{"type":"string","description":"wifi identity"},"keyMgmt":{"type":"string","description":"wifi keyMgmt"},"eap":{"type":"string","description":"wifi eap"},"ip":{"type":"string","description":"wifi ip"},"latitude":{"type":"number","description":"wifi latitude"},"longitude":{"type":"number","description":"wifi longitude"},"connectable":{"type":"boolean","description":"wifi connectable"},"tryTime":{"type":"number","description":"try timestamp"},"accuracy":{"type":"string","description":"wifi accuracy"},"sharedable":{"type":"boolean","description":"wifi sharedable"},"is_root":{"type":"boolean","description":"wifi is_root"},"other_settings":{"type":"string","description":"other setting"}}},"simpleWifiInfo":{"type":"object","properties":{"_id":{"type":"string","description":"db objectId"},"ssid":{"type":"string","description":"wifi ssid"},"bssid":{"type":"string","description":"wifi bssid"},"level":{"type":"integer","description":"wifi level"},"sec_level":{"type":"integer","description":"wifi sec_level"},"capabilities":{"type":"string","description":"wifi capabilities"},"frequency":{"type":"integer","description":"wifi frequency"},"password":{"type":"string","description":"wifi password"},"identity":{"type":"string","description":"wifi identity"},"keyMgmt":{"type":"string","description":"wifi keyMgmt"},"eap":{"type":"string","description":"wifi eap"},"is_root":{"type":"boolean","description":"wifi is_root"},"latitude":{"type":"number","description":"wifi latitude"},"longitude":{"type":"number","description":"wifi longitude"}}},"wifiInfoResponse":{"type":"object","properties":{"_id":{"type":"string","description":"db objectId"},"ssid":{"type":"string","description":"wifi ssid"},"bssid":{"type":"string","description":"wifi bssid"},"level":{"type":"integer","description":"wifi level"},"sec_level":{"type":"integer","description":"wifi sec_level"},"capabilities":{"type":"string","description":"wifi capabilities"},"frequency":{"type":"integer","description":"wifi frequency"},"password":{"type":"string","description":"wifi password"},"identity":{"type":"string","description":"wifi identity"},"keyMgmt":{"type":"string","description":"wifi keyMgmt"},"eap":{"type":"string","description":"wifi eap"},"latitude":{"type":"number","description":"wifi latitude"},"longitude":{"type":"number","description":"wifi longitude"},"accuracy":{"type":"string","description":"wifi accuracy"},"is_root":{"type":"boolean","description":"wifi is_root"},"is_hotspot":{"type":"boolean","description":"是否个人wifi热点"},"other_settings":{"type":"string","description":"other setting"},"icon":{"type":"object","properties":{"normal":{"type":"string","description":"标准图Url"},"thumb":{"type":"string","description":"缩略图Url"}}},"country":{"type":"string","description":"国家代码"},"city":{"type":"string","description":"城市代码"}}},"ErrorRtn":{"type":"object","required":["code"],"properties":{"code":{"type":"integer","description":"错误代码","enum":[101,102,103,104,107,108,109,110,111,112,113,114,115,200,404],"default":101},"msg":{"type":"string","description":"错误描述","default":""}}},"OperRtn":{"type":"object","properties":{"err":{"type":"integer","description":"错误代码，操作成功返回0","default":0},"msg":{"type":"string","description":"错误描述信息","default":""}}}}}